requirements.txt
fastapi
uvicorn[standard]
python-multipart
multipart
pathvalidate
concurrent-log-handler
file_read_backwards
colorama
certifi
charset-normalizer==3.2.0
idna==3.4
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
python-dateutil==2.8.2
shared-memory-dict==0.7.2
# development dependencies
black
mypy
pylint
ruff
flake8
Flake8-pyproject
isort
types-python-dateutil
pandas-stubs
pytest

fastapi
uvicorn[standard]
python-multipart
multipart
pathvalidate
concurrent-log-handler
file_read_backwards
colorama
fastapi-cache2
httpx
